add vet feature and make some refactor

+++ lib/features/shop/ui/widgets/shop_card.dart
@@ -15,7 +15,7 @@ class ShodCard extends StatelessWidget {
   Widget build(BuildContext context) {
     return Container(
       margin: EdgeInsetsDirectional.only(
-        start: sizeConfig.getScreenWidth(AppSize.s16),
+        start: sizeConfig.getScreenWidth(AppSize.s8),
       ),
       decoration: BoxDecoration(
         boxShadow: [customShadow()],
